Move the demo to it's own folder.
