+++ src/demo/app.ts
@@ -1,11 +1,10 @@
-import * as Player from "./player"
-import * as Transport from "./transport/connection"
-
+import { Connection } from "../transport"
+import { Player } from "../player"
 import { Timeline } from "./timeline"
 
 export class App {
-	#transport: Transport.Connection
-	#player: Player.Player // TODO better name
+	#transport: Connection
+	#player: Player
 	#timeline: Timeline
 
 	// DOM elements
@@ -28,13 +27,13 @@ export class App {
 		const params = new URLSearchParams(window.location.search)
 		const url = params.get("url") || "https://localhost:4443"
 
-		this.#transport = new Transport.Connection({
+		this.#transport = new Connection({
 			url,
 			role: "subscriber",
 			fingerprint: url + "/fingerprint",
 		})
 
-		this.#player = new Player.Player(this.#transport, canvas.transferControlToOffscreen())
+		this.#player = new Player(this.#transport, canvas.transferControlToOffscreen())
 		this.#timeline = new Timeline(this.#player, timeline)
 
 		this.#play.addEventListener("click", this.#onPlay.bind(this))
